Add first + last name to signup
Also changed it back to using the form.as_p block token so they can be generated automatically. This
works because the email field is filled in with js, if js is disabled, users can still enter their
email, and it will be still be validated server side. Error messages still are mysteriously
absent...